tests/e2e/pyproject.toml
[project]
name = "verisage-e2e-tests"
version = "0.1.0"
description = "E2E test dependencies for Verisage"
requires-python = ">=3.11"
dependencies = [
    "x402>=0.2.1",
    "eth-account>=0.10.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
